+++ src/KeyVault/KeyVault/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added parameter `ContentType` in `Import-AzKeyVaultCertificate` to support importing pem via certificate string
+* Allowed `DnsName` in `New-AzKeyVaultCertificatePolicy` to accept an empty list [#18954]
+
+## Version 4.6.1
 * Removed the warning messages for MSGraph migration [#18856]
 
 ## Version 4.6.0