+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported to create or update Storage account with Azure Files Active Directory Domain Service Kerberos Authentication
+    -  `New-AzStorageAccount`
+    -  `Set-AzStorageAccount`
+* Supported soft delete DataLake Gen2 item
+    - `Get-AzDataLakeGen2DeletedItem`
+    - `Restore-AzDataLakeGen2DeletedItem`
+
+## Version 4.8.0
 * Added check for storage account sas token is secured with the storage account key.
     -  `New-AzStorageAccountSASToken`
 * Supported Management Policy rule filter BlobIndexMatch